+++ Modules/Intent.Modules.Security.MSAL/release-notes.md
@@ -1,3 +1,7 @@
+### Version 4.2.13
+
+- Improvement: Updated NuGet package versions.
+
 ### Version 4.2.12
 
 - Improvement: Updated NuGet package versions.